requirements.txt
pandas==2.2.3
psycopg2_binary==2.9.10
passlib==1.7.4
bcrypt==4.3.0
plotly==6.0.1
streamlit_cookies_controller==0.0.4
streamlit==1.32.0
streamlit_calendar==1.2.1
streamlit_free_text_select==0.2.0
